♻️ refactor : 명함 상세 mutation 경로 변경
