+++ src/features/card-detail/ui/bottomSheet.tsx
@@ -7,7 +7,7 @@ import BottomModalTitle from '@/shared/ui/bottomModal/bottomModalTitle';
 import { MemoInput } from '@/shared/ui/bottomModal/memoInput';
 import CommonDialog from '@/shared/ui/dialog/commonDialog';
 
-import { useDeleteMyCardMutation, useDeleteReceivedCardMutation } from '../hooks/query/useCardDetailQuery';
+import { useDeleteReceivedCardMutation, useDeleteMyCardMutation } from '../hooks/mutation/useCardDeleteMutation';
 
 type BottomSheetProps = {
   mode: boolean;
